+++ packages/components/src/local/organisms/setting-channels/channels/custom.tsx
@@ -6,19 +6,18 @@ import TableCell from '@material-ui/core/TableCell'
 import TableContainer from '@material-ui/core/TableContainer'
 import TableHead from '@material-ui/core/TableHead'
 import TableRow from '@material-ui/core/TableRow'
+import { ForceData, TemplateBody } from '@serge/custom-types'
 import { ChannelCustom } from '@serge/custom-types/channel-data'
-import { ParticipantCustom } from '@serge/custom-types/participant'
+import { ParticipantCustom, ParticipantTemplate } from '@serge/custom-types/participant'
 import cx from 'classnames'
 import React, { useEffect, useState } from 'react'
 import Confirm from '../../../atoms/confirm'
 import FormGroup from '../../../atoms/form-group-shadow'
-import EditableRow, { Item as RowItem, Option } from '../../../molecules/editable-row'
-import createParticipant from '../helpers/createParticipant'
+import EditableRow, { EDITABLE_SELECT_ITEM, Item as RowItem, Option } from '../../../molecules/editable-row'
+import { SelectItem } from '../../../molecules/editable-row/types/props'
 import { defaultParticipantCustom } from '../helpers/defaultParticipant'
-import generateRowItemsCustom from '../helpers/generateRowItemsCustom'
-import rowToParticipantCustom from '../helpers/rowToParticipantCustom'
 import styles from '../styles.module.scss'
-import { CustomChannelProps } from '../types/props'
+import { CustomChannelProps, Role } from '../types/props'
 
 export const CustomChannel: React.FC<CustomChannelProps> = ({
   channel,
@@ -53,8 +52,85 @@ export const CustomChannel: React.FC<CustomChannelProps> = ({
       setLocalChannelUpdates(newChannel)
     }
 
+    const rowToParticipantCustom = (forces: ForceData[], nextItems: RowItem[], participant: ParticipantCustom): ParticipantCustom => {
+      const [force, access, templateOrPermission] = nextItems.filter(item => item.type === EDITABLE_SELECT_ITEM) as SelectItem[]
+      const selectedForce = forces[force.active ? force.active[0] : 0]
+      const roles: Array<Role['roleId']> = access.active ? access.active.map((key: number) => (selectedForce.roles[key].roleId)) : []
+      const templates: ParticipantTemplate[] = templateOrPermission.active ? templateOrPermission.active.map((key: number) => {
+        const { _id, title } = messageTemplatesOptions[key].value as TemplateBody
+        return { _id, title }
+      }) : []
+
+      return {
+        ...participant,
+        force: selectedForce.name,
+        forceUniqid: selectedForce.uniqid,
+        roles,
+        templates
+      }
+    }
+
+    const generateRowItemsCustom = (templatesOptions: Option[], forces: ForceData[], nextParticipant: ParticipantCustom): RowItem[] => {
+      let forceSelected: number[] = [0]
+      let roleOptions: Option[] = []
+      const additionalFields: RowItem[] = []
+
+      if (nextParticipant.forceUniqid) {
+        const forceIndex = forces.findIndex(force => force.uniqid === nextParticipant.forceUniqid)
+        if (forceIndex !== -1) {
+          roleOptions = forces[forceIndex].roles.map((role): Option => ({
+            name: role.name,
+            uniqid: role.name,
+            value: role
+          }))
+          forceSelected = [forceIndex]
+        }
+      }
+
+      const partRoles: string[] = nextParticipant.roles
+      const activeRoles: number[] = partRoles ? partRoles.map(role => {
+        return roleOptions.findIndex(option => option.value.roleId === role)
+      }).filter(active => active !== -1) : []
+
+      let activeTemplates: number[] = []
+
+      if (nextParticipant.templates && nextParticipant.templates.length) {
+        activeTemplates = nextParticipant.templates.map(template => {
+          return templatesOptions.findIndex(option => option.uniqid === template._id)
+        }).filter(active => active !== -1)
+      }
+
+      additionalFields.push({
+        active: activeTemplates,
+        emptyTitle: 'Chat if empty',
+        multiple: true,
+        options: templatesOptions,
+        uniqid: 'templates',
+        type: EDITABLE_SELECT_ITEM
+      })
+
+      return [
+        {
+          active: forceSelected,
+          multiple: false,
+          options: forces,
+          uniqid: 'forces',
+          type: EDITABLE_SELECT_ITEM
+        },
+        {
+          active: activeRoles,
+          emptyTitle: 'All roles',
+          multiple: true,
+          options: roleOptions,
+          uniqid: 'access',
+          type: EDITABLE_SELECT_ITEM
+        },
+        ...additionalFields
+      ]
+    }
+
     const handleChangeRow = (nextItems: RowItem[], participant: ParticipantCustom): RowItem[] => {
-      const nextParticipant = rowToParticipantCustom(messageTemplatesOptions, forces, nextItems, participant)
+      const nextParticipant = rowToParticipantCustom(forces, nextItems, participant)
       return generateRowItemsCustom(messageTemplatesOptions, forces, nextParticipant)
     }
 
@@ -62,7 +138,7 @@ export const CustomChannel: React.FC<CustomChannelProps> = ({
       if (localChannelUpdates) {
         handleSaveRows([
           ...localChannelUpdates.participants,
-          createParticipant(messageTemplatesOptions, forces, rowItems, localChannelUpdates.channelType) as ParticipantCustom
+          rowToParticipantCustom(forces, rowItems, defaultParticipantCustom)
         ])
       } else {
         console.warn('Can`t create new participant, no current channel')
@@ -77,7 +153,7 @@ export const CustomChannel: React.FC<CustomChannelProps> = ({
             return
           }
           const nextParticipants = [...data.participants]
-          nextParticipants[pKey] = rowToParticipantCustom(messageTemplatesOptions, forces, row, participant)
+          nextParticipants[pKey] = rowToParticipantCustom(forces, row, participant)
           handleSaveRows(nextParticipants)
         }
